Add files via upload
This is the final project that was done in my B. Tech final Year. The idea behind it is a database management system is developed for the hospitals to maintain all the data efficiently that can be accessed by patients, doctors and administrators via a common website. This system facilitates the patients to consult the doctors online and to provide online prescriptions via a chat box which facilitates both the patient and doctor to communicate. And this website facilitates the patients to check their health condition based on the symptoms, patients can also consult doctors at their convenience by sitting at home.